mmc: host: fix issue with new pm code
With the new suspend mechanism in place, mmc host is suspended before
the mmc core and that results in the core layer trying to put the eMMC
card to sleep but since the host is already suspended at that point,
suspend/resume fails then.

Soon after the suspend/resume is retried but the same thing happens
causing the system to never suspend. To fix this, never let the mmc
host suspend before the mmc core.

Change-Id: I1a38bfaa188db937714b1a519a912c60830c65f9
Signed-off-by: Krishna Konda <kkonda@codeaurora.org>
Signed-off-by: Sujit Reddy Thumma <sthumma@codeaurora.org>

Conflicts:
	drivers/mmc/core/host.c